# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: VectorVest Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      security: [{ cookieAuth: [] }]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      summary: Add or toggle a reaction on a message
      security: [{ cookieAuth: [] }]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
  /messages/{messageId}/reactions/{emoji}:
    delete:
      operationId: removeReaction
      tags: [messages]
      summary: Remove a specific reaction from a message
      security: [{ cookieAuth: [] }]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
        - name: emoji
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users/online:
    get:
      operationId: getOnlineUsers
      tags: [users]
      summary: Get online users
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: Online users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineUsersResponse"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    Message:
      type: object
      required: [id, channelId, userId, username, content, createdAt, reactions, isPinned]
      properties:
        id:
          type: integer
        channelId:
          type: string
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionGroup"
        isPinned:
          type: boolean
...
    ReactionGroup:
      type: object
      required: [emoji, count, userReacted]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
    UserProfile:
      type: object
      required: [userId, username, role, membershipStatus]
      properties:
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        membershipStatus:
          type: string
          enum: [pending, verified, rejected]
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 2
    OnlineUsersResponse:
      type: object
      required: [count, users]
      properties:
        count:
          type: integer
        users:
          type: array
          items:
            $ref: "#/components/schemas/OnlineUser"
    OnlineUser:
      type: object
      required: [userId, username]
      properties:
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
...